+++ packages/lit-html/src/directives/repeat.ts
@@ -21,7 +21,7 @@ export type ItemTemplate<T> = (item: T, index: number) => unknown;
 // of an array (used to lazily generate `newKeyToIndexMap` and
 // `oldKeyToIndexMap`)
 const generateMap = (list: unknown[], start: number, end: number) => {
-  const map = new Map();
+  const map = new Map<unknown, number>();
   for (let i = start; i <= end; i++) {
     map.set(list[i], i);
   }
@@ -63,6 +63,12 @@ class RepeatDirective extends Directive {
     };
   }
 
+  render<T>(items: Iterable<T>, template: ItemTemplate<T>): Array<unknown>;
+  render<T>(
+    items: Iterable<T>,
+    keyFn: KeyFn<T> | ItemTemplate<T>,
+    template: ItemTemplate<T>
+  ): Array<unknown>;
   render<T>(
     items: Iterable<T>,
     keyFnOrTemplate: KeyFn<T> | ItemTemplate<T>,
@@ -417,11 +423,19 @@ class RepeatDirective extends Directive {
   }
 }
 
-export type RepeatDirectiveFn = <T>(
-  items: Iterable<T>,
-  keyFnOrTemplate: KeyFn<T> | ItemTemplate<T>,
-  template?: ItemTemplate<T>
-) => unknown;
+export interface RepeatDirectiveFn {
+  <T>(
+    items: Iterable<T>,
+    keyFnOrTemplate: KeyFn<T> | ItemTemplate<T>,
+    template?: ItemTemplate<T>
+  ): unknown;
+  <T>(items: Iterable<T>, template: ItemTemplate<T>): unknown;
+  <T>(
+    items: Iterable<T>,
+    keyFn: KeyFn<T> | ItemTemplate<T>,
+    template: ItemTemplate<T>
+  ): unknown;
+}
 
 /**
  * A directive that repeats a series of values (usually `TemplateResults`)
@@ -434,7 +448,7 @@ export type RepeatDirectiveFn = <T>(
  * will always be created for new keys). This is generally the most efficient
  * way to use `repeat` since it performs minimum unnecessary work for insertions
  * and removals.
- * 
+ *
  * The `keyFn` takes two parameters, the item and its index, and returns a unique key value.
  *
  * ```js
@@ -450,3 +464,9 @@ export type RepeatDirectiveFn = <T>(
  * items to values, and DOM will be reused against potentially different items.
  */
 export const repeat = directive(RepeatDirective) as RepeatDirectiveFn;
+
+/**
+ * The type of the class that powers this directive. Necessary for naming the
+ * directive's return type.
+ */
+export type {RepeatDirective};