+++ packages/lit-html/src/test/static_test.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 import {render} from '../lit-html.js';
-import {html, unsafeStatic} from '../static.js';
+import {html, literal, unsafeStatic} from '../static.js';
 import {assert} from '@esm-bundle/chai';
 import {stripExpressionComments} from './test-utils/strip-markers.js';
 
@@ -16,13 +16,13 @@ suite('static', () => {
   });
 
   test('static text binding', () => {
-    render(html`${unsafeStatic('<p>Hello</p>')}`, container);
+    render(html`${literal`<p>Hello</p>`}`, container);
     // If this were a dynamic binding, the tags would be escaped
     assert.equal(stripExpressionComments(container.innerHTML), '<p>Hello</p>');
   });
 
   test('static attribute binding', () => {
-    render(html`<div class="${unsafeStatic('cool')}"></div>`, container);
+    render(html`<div class="${literal`cool`}"></div>`, container);
     assert.equal(
       stripExpressionComments(container.innerHTML),
       '<div class="cool"></div>'
@@ -32,19 +32,19 @@ suite('static', () => {
   });
 
   test('static tag binding', () => {
-    const tagName = unsafeStatic('div');
+    const tagName = literal`div`;
     render(html`<${tagName}>${'A'}</${tagName}>`, container);
     assert.equal(stripExpressionComments(container.innerHTML), '<div>A</div>');
   });
 
   test('static attribute name binding', () => {
-    render(html`<div ${unsafeStatic('foo')}="${'bar'}"></div>`, container);
+    render(html`<div ${literal`foo`}="${'bar'}"></div>`, container);
     assert.equal(
       stripExpressionComments(container.innerHTML),
       '<div foo="bar"></div>'
     );
 
-    render(html`<div x-${unsafeStatic('foo')}="${'bar'}"></div>`, container);
+    render(html`<div x-${literal`foo`}="${'bar'}"></div>`, container);
     assert.equal(
       stripExpressionComments(container.innerHTML),
       '<div x-foo="bar"></div>'
@@ -52,10 +52,7 @@ suite('static', () => {
   });
 
   test('static attribute name binding', () => {
-    render(
-      html`<div ${unsafeStatic('foo')}="${unsafeStatic('bar')}"></div>`,
-      container
-    );
+    render(html`<div ${literal`foo`}="${literal`bar`}"></div>`, container);
     assert.equal(
       stripExpressionComments(container.innerHTML),
       '<div foo="bar"></div>'
@@ -63,14 +60,14 @@ suite('static', () => {
   });
 
   test('dynamic binding after static text binding', () => {
-    render(html`${unsafeStatic('<p>Hello</p>')}${'<p>World</p>'}`, container);
+    render(html`${literal`<p>Hello</p>`}${'<p>World</p>'}`, container);
     assert.equal(
       stripExpressionComments(container.innerHTML),
       '<p>Hello</p>&lt;p&gt;World&lt;/p&gt;'
     );
 
     // Make sure `null` is handled
-    render(html`${unsafeStatic('<p>Hello</p>')}${null}`, container);
+    render(html`${literal`<p>Hello</p>`}${null}`, container);
     assert.equal(stripExpressionComments(container.innerHTML), '<p>Hello</p>');
   });
 
@@ -125,5 +122,46 @@ suite('static', () => {
     // Static values do not have any caching behavior. Re-rendering with a
     // previously used value does not restore static DOM
     assert.notStrictEqual(div3, div);
+  });
+
+  test('interpolating statics into statics', () => {
+    const start = literal`<${literal`sp${literal`an`}`}>`;
+    const end = literal`</${unsafeStatic('span')}>`;
+    render(html`<div>a${start}b${end}c</div>`, container);
+    assert.equal(
+      stripExpressionComments(container.innerHTML),
+      '<div>a<span>b</span>c</div>'
+    );
+  });
+
+  test('interpolating non-statics into statics throws', () => {
+    assert.throws(() => {
+      literal`a${literal`bar`}b${'shouldthrow'}`;
+    });
+  });
+
+  suite('unsafe', () => {
+    test('static tag binding', () => {
+      const tagName = unsafeStatic('div');
+      render(html`<${tagName}>${'A'}</${tagName}>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>A</div>'
+      );
+    });
+
+    test('static attribute name binding', () => {
+      render(html`<div ${unsafeStatic('foo')}="${'bar'}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="bar"></div>'
+      );
+
+      render(html`<div x-${unsafeStatic('foo')}="${'bar'}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div x-foo="bar"></div>'
+      );
+    });
   });
 });